% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{print.ppcs_track}
\alias{print.ppcs_track}
\title{Print method for ppcs_track objects}
\usage{
\method{print}{ppcs_track}(x, ...)
}
\arguments{
\item{x}{A ppcs_track object.}

\item{...}{Further arguments (unused).}
}
\description{
Print method for ppcs_track objects
}
